+++ sdk/search/search-documents/src/serviceModels.ts
@@ -3574,6 +3574,8 @@ export interface ListKnowledgeSourcesOptions extends OperationOptions {
 export interface CreateKnowledgeSourceOptions extends OperationOptions {}
 export interface GetKnowledgeSourceStatusOptions extends OperationOptions {}
 export interface UploadKnowledgeSourceFileOptions extends OperationOptions {}
+export interface UploadKnowledgeSourceFileMultipartOptions extends OperationOptions {}
+export interface UpdateKnowledgeSourceFileOptions extends OperationOptions {}
 export interface ListKnowledgeSourceFilesOptions extends OperationOptions {}
 export interface DeleteKnowledgeSourceFileOptions extends OperationOptions {}